refactor: Test 코드 분리
